requirements.txt
flask
mistralai
pandas
numpy
python-dotenv
scikit-learn
